pizec no sleep

+++ include/cpu/core.hh
@@ -16,7 +16,7 @@ namespace simlinx {
 
   public:
     void run(reg_t pc);
-    void decode(uint64_t decodedBits, ISA::BasedInstruction &decodedInstr);
+    void decode(uint32_t decodedBits, ISA::BasedInstruction &decodedInstr);
     Core(RAM &ram) : mem(ram){};
   };
 } // namespace simlinx
